Add security measures and fix cross-database issues
Security:
- Add allocation funds validation to prevent over-allocation
- Add donation security for organizers (prevent goal reduction below collected)
- Prevent deletion of campaigns with donations or allocations
- Add client-side and server-side validation for goal amount

Fixes:
- Fix event detail view for public/donor users (cross-database)
- Fix event browse page volunteer count (cross-database safe)
- Fix campaign days left decimal display (cast to int)
- Fix event days until decimal display

Features:
- Add dropdown calendar selector to volunteer schedule page
- Add month/year dropdowns for quick navigation

+++ resources/views/donation-management/public-user/campaign-detail.blade.php
@@ -83,7 +83,7 @@
                             <path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 8v4l3 3m6-3a9 9 0 11-18 0 9 9 0 0118 0z"></path>
                         </svg>
                         @php
-                            $daysLeft = now()->diffInDays($campaign->End_Date, false);
+                            $daysLeft = (int) now()->diffInDays($campaign->End_Date, false);
                             $isExpired = $daysLeft < 0;
                             $daysLeft = abs($daysLeft);
                         @endphp
